chore: SessionId를 SessionNumber로 수정 #24

+++ src/main/java/eightplusone/bit/fit/domain/session/controller/SessionWebSocketController.java
@@ -20,7 +20,7 @@ public class SessionWebSocketController {
 
 	@Scheduled(fixedRate = 30000)
 	public void broadcastSessionUpdate() {
-		Map<Long, Map<String, Object>> sessionData = sessionService.getUpdatedSessionData();
+		Map<Integer, Map<String, Object>> sessionData = sessionService.getUpdatedSessionData();
 		log.info("send congestion data : {}", sessionData.toString());
 		messagingTemplate.convertAndSend("/sub/ws-room", sessionData);
 	}
